add constaints

+++ database/migrations/2025_12_22_163234_add_index_to_xhprof_table.php
@@ -13,6 +13,10 @@ class AddIndexToXHProfTable extends Migration
 	 */
 	public function up()
 	{
+		if (DB::connection()->getDriverName() !== 'mysql') {
+			return;
+		}
+
 		Schema::table('details', function (Blueprint $table) {
 			$table->dropColumn('idcount');
 			$table->char('id', 64)->primary()->change();
@@ -26,6 +30,10 @@ class AddIndexToXHProfTable extends Migration
 	 */
 	public function down()
 	{
+		if (DB::connection()->getDriverName() !== 'mysql') {
+			return;
+		}
+
 		Schema::table('details', function (Blueprint $table) {
 			$table->id('idcount');
 			$table->char('id', 64)->change();
